Factor out _current.py.
